Move seeding logic to the base class
